pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "adamo jordan", email = "adamo.figueroa@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "docling>=2.68.0",
    "fastapi>=0.128.0",
    "google-genai>=1.59.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "transformers>=4.45.0",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
